Added ProgressBar

+++ Assets/Scripts/ManufactoringScripts/ProgressBar.cs
@@ -7,12 +7,12 @@ public class ProgressBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxTime(int time)
+    public void SetMaxTime(float time)
     {
         slider.maxValue = time;
         slider.value = time;
     }
-   public void SetTime(int time)
+   public void SetTime(float time)
     {
         slider.value = time;
     }
